fix: BigDecimal 이 지수표기법으로 나오는 문제 해결 #61

+++ common/src/main/java/com/msa/common/config/MoneyDeserializer.java
@@ -6,13 +6,14 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.msa.common.vo.Money;
 import java.io.IOException;
+import java.math.BigDecimal;
 
 public class MoneyDeserializer extends JsonDeserializer<Money> {
 
     @Override
     public Money deserialize(JsonParser jsonParser, DeserializationContext deserializationContext)
         throws IOException, JacksonException {
-        int value = jsonParser.getIntValue();
+        BigDecimal value = jsonParser.getDecimalValue();
         return new Money(value);
     }
 }
